(UX) Post-update CSS tweaks: disable fixed height and scrollbar
Scrollbar around the entire calendar became a default when we updated
the "fullcalendar" library. This restores the original behavior
(no scrollbar, long events would cause the calendar to be more tall)

Additionally, we increase "minimal height" of an empty calendar cell
from 4em to 6em, because 4em looks too small in situations where the
calendar has been extended to be higher than the default height.