Experimental: add FullCalendar v5 as an option
FullCalendar v2 works really well (and remains the default for now),
but it is quite ancient and will eventually need to be replaced.

Eventually v5 should become the default.
Before that, its CSS must be polished to look better,
and some feedback on it must be gathered from users of JsCalendar
who have complex setups and may notice issues that are not evident
from synthetic testcases.