(feature) Don't exclude the last day of events with start/end date
See issue #8.

FullCalendar library (which we use) doesn't count the end date as a part
of the event, so 2022/04/28:2022/04/30 wasn't showing April 30.

This is confusing/unexpected behavior for editors/readers,
because 2022/04/30 will be in the page title,
but the calendar won't have anything on this day.

Including the end date makes a lot of sense, and is a new default.

New parameter excludeEndDate=1 restores the previous behavior
(in case somebody needs it).